requirements.txt
audioop-lts==0.2.1
audioread==3.0.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
colorama==0.4.6
decorator==5.2.1
ffmpeg-python==0.2.0
filelock==3.13.1
fsspec==2024.6.1
future==1.0.0
idna==3.10
Jinja2==3.1.4
joblib==1.5.1
lazy_loader==0.4
librosa==0.11.0
llvmlite==0.44.0
MarkupSafe==2.1.5
more-itertools==10.7.0
mpmath==1.3.0
msgpack==1.1.1
networkx==3.3
numba==0.61.2
numpy==2.1.2
openai-whisper==20250625
opencv-python==4.11.0.86
packaging==25.0
pillow==11.0.0
platformdirs==4.3.8
pooch==1.8.2
PyAudio==0.2.14
pycparser==2.22
PyQt6==6.9.1
PyQt6-Qt6==6.9.1
PyQt6_sip==13.10.2
regex==2024.11.6
requests==2.32.4
Resemblyzer==0.1.4
scikit-learn==1.7.0
scipy==1.16.0
setuptools==70.2.0
SoundCard==0.4.4
sounddevice==0.5.2
soundfile==0.13.1
soxr==0.5.0.post1
standard-aifc==3.13.0
standard-chunk==3.13.0
standard-sunau==3.13.0
sympy==1.13.3
threadpoolctl==3.6.0
tiktoken==0.9.0
torch==2.7.1+cu118
torchaudio==2.7.1+cu118
torchvision==0.22.1+cu118
tqdm==4.67.1
typing==3.7.4.3
typing_extensions==4.12.2
urllib3==2.5.0
webrtcvad==2.0.10
